# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kampala Bakery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wholesale/customers:
    get:
      operationId: listWholesaleCustomers
      tags: [wholesale]
      summary: List business customers
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WholesaleCustomer"
    post:
      operationId: createWholesaleCustomer
      tags: [wholesale]
      summary: Add a business customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWholesaleCustomerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WholesaleCustomer"
  /wholesale/supplies:
    get:
      operationId: listWholesaleSupplies
      tags: [wholesale]
      summary: List supply records
      responses:
        "200":
          description: Supply records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WholesaleSupply"
    post:
      operationId: createWholesaleSupply
      tags: [wholesale]
      summary: Record a supply to a business customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWholesaleSupplyBody"
      responses:
        "201":
          description: Supply recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WholesaleSupply"
  /wholesale/supplies/{id}/payment:
    put:
      operationId: updateSupplyPaymentStatus
      tags: [wholesale]
      summary: Update payment status for a supply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WholesaleSupply"
...
components:
  schemas:
...
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required: [productId, productName, quantity, unitPrice, subtotal]
...
    SaleItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
...
    WholesaleCustomer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        location:
          type: string
        totalOutstanding:
          type: number
      required: [id, name, contactPerson, phone, location, totalOutstanding]
    CreateWholesaleCustomerBody:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        location:
          type: string
      required: [name, contactPerson, phone, location]
    WholesaleSupply:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        totalAmount:
          type: number
        paymentStatus:
          type: string
          enum: [paid, unpaid, credit]
        amountPaid:
          type: number
        suppliedAt:
          type: string
        notes:
          type: string
          nullable: true
      required: [id, customerId, customerName, items, totalAmount, paymentStatus, amountPaid, suppliedAt]
    CreateWholesaleSupplyBody:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
        paymentStatus:
          type: string
          enum: [paid, unpaid, credit]
        amountPaid:
          type: number
        notes:
          type: string
      required: [customerId, items, paymentStatus, amountPaid]
    UpdatePaymentStatusBody:
      type: object
      properties:
        paymentStatus:
          type: string
          enum: [paid, unpaid, credit]
        amountPaid:
          type: number
      required: [paymentStatus, amountPaid]
...